Modify usuario.py controller

+++ controllers/usuario.py
@@ -1,3 +1,6 @@
+import hashlib
+import ldap
+
 def login():
     form = SQLFORM.factory(
         Field('username', requires=IS_NOT_EMPTY()),
@@ -7,18 +10,7 @@ def login():
     if form.process(hideerror=True).accepted:
         username = form.vars.username
         password = form.vars.password
-        user = auth.login_bare(username, password)
-
-        if user:
-            edicion_actual = db(db.edicion.activa == True).select().first()
-
-            if not edicion_actual:
-                plugin_toastr_message_config('success', T(
-                    'No hay ediciones activas'))
-            else:
-                plugin_toastr_message_config('success', T(
-                    'Bienvenido al sistema'))
-
+        if check_credentials(username, password):
             redirect(auth.settings.login_next)
         else:
             form.errors.credenciales = T('Credenciales incorrectas')
@@ -31,9 +23,32 @@ def login():
     return dict(form=form)
 
 
+def check_credentials(username, password):
+    LDAP_SERVER = 'ldap://10.18.1.2'
+    LDAP_USERNAME = '%s@unica.cu' % username
+    LDAP_PASSWORD = password
+    base_dn = 'DC=unica,DC=cu'
+    ldap_filter = 'userPrincipalName=%s@unica.cu' % username
+    attrs = ['memberOf']
+    try:
+        ldap_client = ldap.initialize(LDAP_SERVER)
+        ldap_client.set_option(ldap.OPT_REFERRALS, 0)
+        ldap_client.simple_bind_s(LDAP_USERNAME, LDAP_PASSWORD)
+        ab = str(ldap_client.search_s(base_dn,
+                                      ldap.SCOPE_SUBTREE, ldap_filter, attrs)[0][1]['memberOf'])
+        ldap_client.unbind()
+        session.user = username
+        return True
+    except ldap.INVALID_CREDENTIALS:
+        return False
+    except ldap.SERVER_DOWN:
+        session.errorl = "Error de conexión con el servidor"
+        redirect(URL('usuario', 'login'))
+        return False
+
 def register():
     usuarios = db(db.auth_user.id > 0)
-    require_username = [IS_NOT_EMPTY(), IS_NOT_IN_DB(
+    require_username = [IS_NOT_EMPTY(), IS_NOT_IN_DB
         usuarios, "auth_user.username"), IS_LENGTH(20)]
     require_email = [IS_NOT_EMPTY(), IS_EMAIL(
     ), IS_NOT_IN_DB(usuarios, "auth_user.email")]
@@ -52,28 +67,33 @@ def register():
 
                            )
 
-    if form.process(hideerror=True).accepted:
-        user_id = db.auth_user.insert(**form.vars)
-        group_id = db(db.auth_group.role == "Estudiante").select().first().id
-        db.auth_membership.insert(
-            user_id=user_id, group_id=group_id)
-        plugin_toastr_message_config('success', T(
-            'El registro se realizó correctamente'))
-        redirect(URL("usuario", "login"))
-
+    if form.process().accepted:
+        if register_user_ldap(form.vars.username, form.vars.password):
+            user_id = db.auth_user.insert(**form.vars)
+            group_id = db(db.auth_group.role == "Estudiante").select().first().id
+            db.auth_membership.insert(
+                user_id=user_id, group_id=group_id)
+            plugin_toastr_message_config('success', T(
+                'El registro se realizó correctamente'))
+            redirect(URL("usuario", "login"))
+        else:
+            plugin_toastr_message_config('error', T(
+                'No se pudo registrar el usuario LDAP'))
     elif form.errors:
         plugin_toastr_message_config('error', T(
             'El formulario contiene errores'))
 
     return dict(form=form)
 
+def register_user_ldap(username, password):
+    # Implementa la lógica para registrar usuarios en el servidor LDAP
+    # Retorna True si el registro es exitoso, False en caso contrario
+    return False  # Implementa la lógica adecuada aquí
 
 @auth.requires_login()
 def logout():
-    # Cierra la sesión del usuario dado
     auth.logout()
     return dict()
-
 
 @auth.requires_login()
 def perfil():
@@ -88,17 +108,26 @@ def perfil():
                                IS_EQUAL_TO(request.vars.password)]),
                            )
 
-    if form.process(hideerror=True).accepted:
-        db(db.auth_user.id == usuario.id).update(**form.vars)
-        plugin_toastr_message_config('success', T(
-            'Contraseña actualizada correctamente'))
-        redirect(URL("perfil", args=usuario.id))
+    if form.process().accepted:
+        # Actualiza la contraseña en el servidor LDAP
+        if update_password_ldap(usuario.username, form.vars.password):
+            db(db.auth_user.id == usuario.id).update(**form.vars)
+            plugin_toastr_message_config('success', T(
+                'Contraseña actualizada correctamente'))
+            redirect(URL("perfil"))
+        else:
+            plugin_toastr_message_config('error', T(
+                'No se pudo actualizar la contraseña en el servidor LDAP'))
     elif form.errors:
         plugin_toastr_message_config('error', T(
             'El formulario contiene errores'))
 
     return dict(usuario=usuario, rol=rol, form=form)
 
+def update_password_ldap(username, password):
+    # Implementa la lógica para actualizar la contraseña en el servidor LDAP
+    # Retorna True si la actualización es exitosa, False en caso contrario
+    return False  # Implementa la lógica adecuada aquí
 
 @auth.requires_membership('Administrador')
 def manage():
@@ -106,7 +135,6 @@ def manage():
         db.auth_user.username != "admin") &
         (db.auth_membership.user_id == db.auth_user.id)).select(db.auth_user.ALL, db.auth_membership.ALL)
     return dict(usuarios=usuarios)
-
 
 @auth.requires(auth.has_membership('Administrador') or auth.has_membership('Coordinador'))
 def detalles():
@@ -120,7 +148,6 @@ def detalles():
              usuario.id).select().first().group_id.role
 
     return dict(usuario=usuario, rol=rol)
-
 
 @auth.requires_membership('Administrador')
 def crear():
@@ -155,24 +182,26 @@ def crear():
             plugin_toastr_message_config('error', T(
                 'El formulario contiene errores'))
         else:
-            user_id = db.auth_user.insert(**form.vars)
-            db.auth_membership.insert(
-                user_id=user_id, group_id=form.vars.rol)
+            if register_user_ldap(form.vars.username, form.vars.password):
+                user_id = db.auth_user.insert(**form.vars)
+                db.auth_membership.insert(
+                    user_id=user_id, group_id=form.vars.rol)
 
-            if form.vars.rol == 2:
-                db.coordinador_maestria.insert(
-                    id_maestria=form.vars.maestria, id_user=user_id)
+                if form.vars.rol == 2:
+                    db.coordinador_maestria.insert(
+                        id_maestria=form.vars.maestria, id_user=user_id)
 
-            plugin_toastr_message_config('success', T(
-                'Usuario creado correctamente'))
-            redirect(URL("usuario", "manage"))
-
+                plugin_toastr_message_config('success', T(
+                    'Usuario creado correctamente'))
+                redirect(URL("usuario", "manage"))
+            else:
+                plugin_toastr_message_config('error', T(
+                    'No se pudo crear el usuario LDAP'))
     elif form.errors:
         plugin_toastr_message_config('error', T(
             'El formulario contiene errores'))
 
     return dict(form=form)
-
 
 @auth.requires_membership("Administrador")
 def editar():
@@ -242,7 +271,6 @@ def editar():
 
     return dict(form=form, registro=registro)
 
-
 @auth.requires_membership("Administrador")
 def eliminar():
     if not request.args(0):
@@ -253,13 +281,21 @@ def eliminar():
     x = registro.id
 
     if not registro.username == "admin":
-        db(db.auth_user.id == x).delete()
-        plugin_toastr_message_config('success', T(
-            'Usuario eliminado correctamente'))
+        if delete_user_ldap(registro.username):
+            db(db.auth_user.id == x).delete()
+            plugin_toastr_message_config('success', T(
+                'Usuario eliminado correctamente'))
+        else:
+            plugin_toastr_message_config('error', T(
+                'No se pudo eliminar el usuario LDAP'))
 
     redirect(URL("manage"))
     return dict()
 
+def delete_user_ldap(username):
+    # Implementa la lógica para eliminar usuarios en el servidor LDAP
+    # Retorna True si la eliminación es exitosa, False en caso contrario
+    return False  # Implementa la lógica adecuada aquí
 
 @auth.requires_membership("Administrador")
 def cambiar_clave():
@@ -275,17 +311,20 @@ def cambiar_clave():
                            )
 
     if form.process(hideerror=True).accepted:
-        db(db.auth_user.id == registro.id).update(**form.vars)
-        plugin_toastr_message_config('success', T(
-            'Contraseña actualizada correctamente'))
-        redirect(URL("detalles", args=registro.id))
+        if update_password_ldap(registro.username, form.vars.password):
+            db(db.auth_user.id == registro.id).update(**form.vars)
+            plugin_toastr_message_config('success', T(
+                'Contraseña actualizada correctamente'))
+            redirect(URL("detalles", args=registro.id))
+        else:
+            plugin_toastr_message_config('error', T(
+                'No se pudo actualizar la contraseña en el servidor LDAP'))
 
     elif form.errors:
         plugin_toastr_message_config('error', T(
             'El formulario contiene errores'))
 
     return dict(form=form, registro=registro)
-
 
 @auth.requires_login()
 def no_autorizado():
